+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## 0.1.0
-- Add IsaacusTextEmbedder and IsaacusDocumentEmbedder components.
+## 0.3.0 
+- Add IsaacusEnricher component.
 
 ## 0.2.0
 - Add IsaacusRanker component.
+
+## 0.1.0
+- Add IsaacusTextEmbedder and IsaacusDocumentEmbedder components.